1. Replace stringbuffer with stringbuilder 2. Update max_width to 110, now I could see the current operation component while using geronimo run to start the server
git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@1144136 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/modules/geronimo-system/src/main/java/org/apache/geronimo/system/main/ProgressBarStartupMonitor.java
@@ -100,7 +100,7 @@ public class ProgressBarStartupMonitor implements StartupMonitor {
                     break;
             }
         }
-        percent += Math.round(90f * (float) progress / (float) total);
+        percent += Math.round(90f * progress / total);
         this.percent = percent;
     }
 
@@ -166,7 +166,7 @@ public class ProgressBarStartupMonitor implements StartupMonitor {
             buf.append(' ');
         }
         buf.append(percent).append("% ");
-        int time = Math.round((float) (System.currentTimeMillis() - started) / 1000f);
+        int time = Math.round((System.currentTimeMillis() - started) / 1000f);
         if (time < 10) {
             buf.append(' ');
         }
